+++ app/app.go
@@ -155,6 +155,10 @@ const (
 	FlagAsyncWAL = "store.memiavl-async-wal"
 	// don't enable zero-copy together with inter-block cache.
 	FlagZeroCopy = "store.memiavl-zero-copy"
+
+	// FlagSaveSnapshotDir saves the state snapshot during state-sync restoration,
+	// it's useful is restoration is slow and user can save the snapshot and try local restoration later.
+	FlagSaveSnapshotDir = "store.save-snapshot-dir"
 )
 
 // this line is used by starport scaffolding # stargate/wasm/app/enabledProposals
@@ -335,6 +339,9 @@ type App struct {
 
 	// module configurator
 	configurator module.Configurator
+
+	// saveSnapshotDir is the directory where the restored snapshot will be saved during the state-sync bootstrap.
+	saveSnapshotDir string
 }
 
 // New returns a reference to an initialized chain.
@@ -398,6 +405,7 @@ func New(
 		keys:              keys,
 		tkeys:             tkeys,
 		memKeys:           memKeys,
+		saveSnapshotDir:   cast.ToString(appOpts.Get(FlagSaveSnapshotDir)),
 	}
 
 	app.ParamsKeeper = initParamsKeeper(appCodec, cdc, keys[paramstypes.StoreKey], tkeys[paramstypes.TStoreKey], skipGravity)